pc_app/requirements.txt
pyserial>=3.5
nvidia-ml-py3>=7.352.0 